requirements.txt
astroid==2.11.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
deprecation==2.1.0
dill==0.3.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
numpy==1.22.3
onboard.client==1.8.0
orjson==3.6.4
packaging==21.3
pandas==1.4.2
platformdirs==2.5.2
pydantic==1.9.0
pylint==2.13.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
